chapter 7 all complate

+++ application/controllers/board.php
@@ -128,45 +128,53 @@ class Board extends CI_Controller
 
     public function write()
     {
-        $this->load->library('form_validation');
-
-        $this->form_validation->set_rules("subject", "제목", "required");
-        $this->form_validation->set_rules("contents", "내용", "required");
-
+        $this->load->helper('alert');
         echo "<meta http-equiv='Content-Type' content='text/html' charset='utf-8' />";
-        // 글쓰기 성공시
-        if ($this->form_validation->run() == true) {
-            // 경고창 헬퍼 로딩
-            $this->load->helper('alert');
-            // 주소 중에서 page세그먼트가 있는지 검사하기 위해 주소를 배열로 변환
-            $uri_array = $this->segment_explode($this->uri->uri_string());
 
-            if (in_array('page', $uri_array)) {
-                $pages = urldecode($this->url_explode($uri_array, 'page'));
-            } else {
-                $pages = 1;
-            }
-            if (! $this->input->post('subject', true) && ! $this->input->post('contents', true)) {
-                alert("비정상적인 접근입니다.", "/codeigniter/board/lists/" . $this->uri->segment(3) . "/page/" . $pages);
-                exit();
-            }
-            $write_data = array(
-                'subject' => $this->input->post('subject', true),
-                'contents' => $this->input->post('contents', true),
-                'table' => $this->uri->segment(3)
-            );
-            $result = $this->board_m->insert_board($write_data);
+        if ($this->session->userdata('logged_in') == true) {
+            $this->load->library('form_validation');
 
-            if ($result) {
-                // 글 작성 성공시 게시물 목록으로
-                alert("입력되었습니다.", "/codeigniter/board/lists/" . $this->uri->segment(3) . "/page/" . $pages);
-                exit();
+            $this->form_validation->set_rules("subject", "제목", "required");
+            $this->form_validation->set_rules("contents", "내용", "required");
+
+            // 글쓰기 성공시
+            if ($this->form_validation->run() == true) {
+                // 경고창 헬퍼 로딩
+
+                // 주소 중에서 page세그먼트가 있는지 검사하기 위해 주소를 배열로 변환
+                $uri_array = $this->segment_explode($this->uri->uri_string());
+
+                if (in_array('page', $uri_array)) {
+                    $pages = urldecode($this->url_explode($uri_array, 'page'));
+                } else {
+                    $pages = 1;
+                }
+                if (! $this->input->post('subject', true) && ! $this->input->post('contents', true)) {
+                    alert("비정상적인 접근입니다.", "/codeigniter/board/lists/" . $this->uri->segment(3) . "/page/" . $pages);
+                    exit();
+                }
+                $write_data = array(
+                    'subject' => $this->input->post('subject', true),
+                    'contents' => $this->input->post('contents', true),
+                    'table' => $this->uri->segment(3),
+                    'user_id' => $this->session->userdata('username')
+                );
+                $result = $this->board_m->insert_board($write_data);
+
+                if ($result) {
+                    // 글 작성 성공시 게시물 목록으로
+                    alert("입력되었습니다.", "/codeigniter/board/lists/" . $this->uri->segment(3) . "/page/" . $pages);
+                    exit();
+                } else {
+                    alert("다시 입력해주세요.", "/codeigniter/board/lists/" . $this->uri->segment(3) . "/page/" . $pages);
+                    exit();
+                }
             } else {
-                alert("다시 입력해주세요.", "/codeigniter/board/lists/" . $this->uri->segment(3) . "/page/" . $pages);
-                exit();
+                $this->load->view("board/write_v");
             }
         } else {
-            $this->load->view("board/write_v");
+            alert("로그인후 이용해주세요.", "/codeigniter/auth/login");
+            exit();
         }
     }
 
@@ -185,24 +193,36 @@ class Board extends CI_Controller
             } else {
                 $pages = 1;
             }
-            if (! $this->input->post('subject', true) && ! $this->input->post('contents', true)) {
-                alert("비정상적인 접근입니다.", "/codeigniter/board/lists/" . $this->uri->segment(3) . "/page/" . $pages);
-                exit();
-            }
-            $write_data = array(
-                'subject' => $this->input->post('subject', true),
-                'contents' => $this->input->post('contents', true),
-                'table' => $this->uri->segment(3),
-                'board_id' => $this->uri->segment(5)
-            );
-            $result = $this->board_m->modify_board($write_data);
 
-            if ($result) {
-                // 글 작성 성공시 게시물 목록으로
-                alert("입력되었습니다.", "/codeigniter/board/lists/" . $this->uri->segment(3) . "/page/" . $pages);
-                exit();
+            if ($this->session->userdata('logged_in') == true) {
+                $write_id = $this->board_m->write_check();
+                if ($write_id->user_id != $this->session->userdata('username')) {
+                    alert("본이이 작성한 글이 아닙니다.", "/codeigniter/board/edit/" . $this->uri->segment(3) . "/board_id/" . $this->uri->segment(5) . "/page/" . $pages);
+                    exit();
+                }
+
+                if (! $this->input->post('subject', true) && ! $this->input->post('contents', true)) {
+                    alert("비정상적인 접근입니다.", "/codeigniter/board/lists/" . $this->uri->segment(3) . "/page/" . $pages);
+                    exit();
+                }
+                $write_data = array(
+                    'subject' => $this->input->post('subject', true),
+                    'contents' => $this->input->post('contents', true),
+                    'table' => $this->uri->segment(3),
+                    'board_id' => $this->uri->segment(5)
+                );
+                $result = $this->board_m->modify_board($write_data);
+
+                if ($result) {
+                    // 글 작성 성공시 게시물 목록으로
+                    alert("입력되었습니다.", "/codeigniter/board/lists/" . $this->uri->segment(3) . "/page/" . $pages);
+                    exit();
+                } else {
+                    alert("다시 입력해주세요.", "/codeigniter/board/lists/" . $this->uri->segment(3) . "/page/" . $pages);
+                    exit();
+                }
             } else {
-                alert("다시 입력해주세요.", "/codeigniter/board/lists/" . $this->uri->segment(3) . "/page/" . $pages);
+                alert("본인이 입력한 글이 아닙니다.", "/codeigniter/auth/login");
                 exit();
             }
         } else {
@@ -218,12 +238,25 @@ class Board extends CI_Controller
     {
         echo "<meta http-equiv='Content-Type' content='text/html' charset='utf-8' />";
         $this->load->helper('alert');
-        $result = $this->board_m->delete_content($this->uri->segment(3), $this->uri->segment(5));
-        if ($result) {
-            alert("삭제되었습니다.", "/codeigniter/board/lists/" . $this->uri->segment(3) . "/page/" . $this->uri->segment(7));
-            exit();
+
+        if ($this->session->userdata('logged_in') == true) {
+            $write_id = $this->board_m->write_check();
+
+            if ($write_id->user_id != $this->session->userdata('username')) {
+                alert("본이이 작성한 글이 아닙니다.", "/codeigniter/board/edit/" . $this->uri->segment(3) . "/board_id/" . $this->uri->segment(5) . "/page/" . $this->uri->segment(7));
+                exit();
+            }
+
+            $result = $this->board_m->delete_content($this->uri->segment(3), $this->uri->segment(5));
+            if ($result) {
+                alert("삭제되었습니다.", "/codeigniter/board/lists/" . $this->uri->segment(3) . "/page/" . $this->uri->segment(7));
+                exit();
+            } else {
+                alert("삭제되었습니다.", "/codeigniter/board/view/" . $this->uri->segment(3) . "/board_id/" . $this->uri->segment(5) . "/page/" . $this->uri->segment(7));
+                exit();
+            }
         } else {
-            alert("삭제되었습니다.", "/codeigniter/board/view/" . $this->uri->segment(3) . "/board_id/" . $this->uri->segment(5) . "/page/" . $this->uri->segment(7));
+            alert("로그인후 이용해주세요.", "/codeigniter/auth/login");
             exit();
         }
     }
